Add files via upload
Added the base implementation for admin/instructor page. Currently it is possible to add files and the files get copied to the `data` directory. The removal part is buggy. The next steps would be to add chatbot configuration functionality, fix file deletion, and configure authentication to switch page to admin/student page.